Add useBlurOnHide
